style: cleaning & finalizing code
